Split up Tooltip module
